dashboard page feature, automatically updates

+++ app/models/bills.js
@@ -5,10 +5,10 @@ module.exports = function(sequelize, DataTypes) {
     user_id: { type: DataTypes.INTEGER(11), allowNull: false },
     email: { type: DataTypes.STRING(64), allowNull: false, validate: {isEmail: true} },
     currency: { type: DataTypes.STRING(64), allowNull: false },
-    currency_amount: { type: DataTypes.INTEGER(11), allowNull: false },
+    currency_amount: { type: DataTypes.DECIMAL(20,8), allowNull: false },
     address: { type: DataTypes.STRING(64), allowNull: false },
     asset_id: { type: DataTypes.INTEGER(11), allowNull: false },
-    asset_amount: { type: DataTypes.INTEGER(11), allowNull: false },
+    asset_amount: { type: DataTypes.DECIMAL(20,8), allowNull: false },
     message: { type: DataTypes.STRING(1024), allowNull: true },
     status: { type: DataTypes.INTEGER(1), allowNull: false },
   }, {
